source/inference/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
apex==0.1
appdirs==1.4.4
ascii-graph==1.5.1
attrs==19.3.0
audioread==2.1.8
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.5
boto3==1.14.0
botocore==1.17.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
codecov==2.1.4
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
coverage==5.1
cryptography==2.9.2
cxxfilt==0.2.1
cycler==0.10.0
cymem==2.0.2
Cython==0.28.4
cytoolz==0.9.0.1
DataProperty==0.49.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.2.9
DLLogger==0.1.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
future==0.18.2
glob2==0.7
grpcio==1.29.0
h5py==2.10.0
html2text==2020.1.16
hypothesis==4.50.8
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.1
inflect==4.1.0
ipdb==0.13.2
ipykernel==5.3.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyter-tensorboard==0.2.0
jupyterlab==1.2.14
jupyterlab-server==1.1.5
jupytext==1.5.0
kiwisolver==1.2.0
libarchive-c==2.9
librosa==0.6.3
llvmlite==0.28.0
lmdb==0.98
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
maskrcnn-benchmark==0.1
matplotlib==3.2.1
mbstrdecoder==1.0.0
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==4.0.2
more-itertools==8.3.0
msgfy==0.1.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.0
nltk==3.5
notebook==6.0.3
numba==0.43.1
numpy==1.18.1
nvidia-dali==0.22.0
nvidia-ml-py3==7.352.0
onnx==1.6.0
onnxruntime==1.3.0
opencv-python==3.4.1.15
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
pathvalidate==2.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow-SIMD==5.3.0.post1
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
progressbar==2.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pybind11==2.5.0
pycocotools==2.0+nv0.4.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycuda==2019.1.2
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyprof==3.1.0
pyrsistent==0.16.0
PySocks==1.7.1
pytablewriter==0.54.0
pytest==5.4.3
pytest-cov==2.9.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python-hostlist==1.20
python-nvd3==0.15.0
python-slugify==4.0.0
pytools==2020.2
pytorch-transformers==1.1.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.5.14
requests==2.23.0
resampy==0.2.2
revtok==0.0.3
ruamel-yaml==0.15.87
s3transfer==0.3.3
sacrebleu==1.2.10
sacremoses==0.0.35
scikit-image==0.15.0
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sox==1.3.7
spacy==2.0.16
Sphinx==3.1.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SSD==0.1
subword-nmt==0.3.3
tabledata==1.1.2
tabulate==0.8.7
tcolorpy==0.0.5
tensorboard==1.15.0+nv
tensorboard-plugin-dlprof==20.6
tensorboardX==2.4
tensorrt==7.1.2.8
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==6.12.1
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
torch==1.6.0a0+9907a3e
torchtext==0.6.0
torchvision==0.7.0a0
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.3
typepy==1.1.1
typing==3.7.4.1
typing-extensions==3.7.4.2
ujson==2.0.3
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.10.11
yacs==0.1.7
zipp==3.1.0
